[장바구니 협업 Step 2] 룩소(우정균) 미션 제출합니다. (#170)
* feat: 병합 충돌 해결

* refactor: fetchCartItemList selector 분리

* refactor: Header 컴포넌트 useCartService 훅 의존성 제거

* refactor: ProductList 컴포넌트 import 순서 변경

* refactor: fetchCartItemList selector 서버가 변경되면 자동으로 주소 변경되도록 변경

* refactor: useCartService 로딩 상태 제거

* refactor: cartLoading 상태 삭제

* refactor: useCheckedCartList 훅 로딩 상태 삭제

* refactor: ProductItem 로딩 상태 제거 및 counter 업데이트 시점 변경

* refactor: CartPage 컴포넌트 EmptyCartView 분리

* refactor: CartList, ProductList 서버 변경 시 cart 전역 상태 초기 값으로 리셋 기능 추가

* refactor: useCheckCartList 훅 서버 변경시 체크 로직 변경

* feat: Header 장바구니 버튼 장바구니 fetching 시 로딩 스피너 보여주도록 설정

* refactor: useFetch 훅 option 받도록 수정

* refactor: EmptyCartView 를 EmptyCartSuspense로 변경 및 적용

* refactor: CartList 카트 페이지 진입 시 카트 리스트 fetching 로직 수정

* refactor: useChangeServer 훅 분리 및 적용

* refactor: CartList useChangeServer 적용

* refactor: useCartService 업데이트 함수 quantity가 0이면 delete하도록 수정

* refactor: ProductItem Counter 상태 가지도록 변경

* refactor: useCheckedCartList 서버 이동시 체크 로직 변경

* feat: useCheckedCartList 서버 이동시 체크 로직 버그 수정

* refactor: ProductItem 장바구니 컨트롤러 컴포넌트 분리

* chore: 리액트 Strict 모드 비활성화

* refactor: useChangeServer -> useResetCartWhenChangeServer로 네이밍 변경

* feat: 로딩 스피너 스타일 변경

* refactor: 로딩 스피너 기본 색상 변경

* feat: 서버 목록에 Mock 서버 추가

* feat: Modal 컴포넌트 구현

* chore: react-icons 라이브러리 설치

* feat: GlobalStyle 기본 글꼴 설정

* feat: Coupon 타입 지정

* feat: CouponItem 컴포넌트 구현 및 스토리북 작성

* refactor: Coupon Type 추가

* refactor: LoadingView 컴포넌트 추가 및 장바구니 페이지, 모달 적용

* feat: mock 서버 쿠폰 발급 api 추가

* feat: CouponItem 스타일 변경

* feat: ServerUtil 쿠폰 관련 url getter 추가

* feat: CouponList 컴포넌트 구현

* feat: CouponList 컴포넌트 스타일 수정 및 CouponItem 컴포넌트 스타일 추가

* feat: 쿠폰 발급 mock api 데이터 개수 추가

* feat: 쿠폰 관련 mock api 모두 추가

* refactor: 쿠폰 url path name 변경

* feat: 모달 오픈시 스크롤 방지 옵션 구현

* refactor: ServerUtil 쿠폰 url path name 변경

* feat: 쿠폰 발급 기능 구현

* refactor: useCheckedCartList 훅 장바구니 체크 리스트 가져오는 함수 추가 및 다른 함수들 네이밍 수정

* fix: mock 핸들러 쿠폰 db 데이터 형식 변경

* feat: PaymentAmount 주문서 모달 오픈 기능 구현

* feat: 주문서 상품 목록 기능 구현

* feat: 주문서 쿠폰 목록 기능 구현

* feat: 주문서 주문 금액 기능 구현

* feat: 주문서 기능 구현 완료

* feat: order api url 추가

* feat: order 타입 추가

* feat: 주문서 주문 기능 구현

* feat: 주문 관련 mock api 구현

* feat: OrderPage 컴포넌트 레이아웃 구성

* refactor: CartPage 파일 이름 오타 수정 및 Error Boundary 추가

* fix: Modal 언마운트 될 때 스크롤 방지 해제 하도록 수정

* fix: 주문 mock api 주문 목록 저장하도록 변경

* fix: useOrderSheet 주문 기능 함수 버그 수정 및 주문한 장바구니 삭제하도록 변경

* feat: Header 주문 목록 버튼 구현

* feat: OrderSheet 쿠폰 목록 부분에 에러 바운더리 설정

* refactor: 에러 바운더리 스타일 변경

* feat: 주문목록 api 데이터 형식 추가

* feat: 주문목록 페이지 구현

* chore: 배포 자동화 파일 브랜치 수정

* refactor: 쿠폰 받기 버튼 스타일 변경 및 모달 z-index 설정

* fix: MSW 주소 에러 재설정

* docs: README 병합 충돌 메시지 제거

* chore: 스토리북 배포 자동화 설정

* docs: README 개행문제 수정 및 배포 링크 추가

* refactor: 사용하지 않는 type export 제거

* refactor: usePaymentAmount 배송비 얻는 로직 함수로 분리

* refactor: constants 상수들 파일 분리 및 라우팅 경로 상수화

* refactor: usePayment getDeliveryFee 함수 추가

* refactor: ServerUtil 객체를 urlConfig로 네이밍 변경

* refactor: console.log 제거

* refactor: fetchCartItemList selector 캐시 옵션 삭제

+++ src/hooks/useCartService.ts
@@ -1,21 +1,17 @@
-import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
+import { useRecoilState, useRecoilValue } from 'recoil';
 import cartState from '../globalState/atoms/cartState';
 import type { Product } from '../types/product';
 import serverNameState from '../globalState/atoms/serverName';
-import ServerUtil from '../utils/ServerUrl';
-import { USER_AUTH_TOKEN } from '../constant';
-import cartLoadingState from '../globalState/atoms/cartLoadingState';
+import urlConfig from '../utils/urlConfig';
+import { USER_AUTH_TOKEN } from '../constant/user';
 
 const useCartService = () => {
   const [cartList, setCartList] = useRecoilState(cartState);
-  const setCartLoading = useSetRecoilState(cartLoadingState);
 
   const serverName = useRecoilValue(serverNameState);
-  const cartItemsUrl = ServerUtil.getCartItemsUrl(serverName);
+  const cartItemsUrl = urlConfig.getCartItemsUrl(serverName);
 
   const fetchCartItem = async () => {
-    setCartLoading(true);
-
     const response = await fetch(cartItemsUrl, {
       headers: {
         Authorization: `Basic ${USER_AUTH_TOKEN}`,
@@ -28,8 +24,6 @@ const useCartService = () => {
 
     const fetchedCartList = await response.json();
     setCartList(fetchedCartList);
-
-    setCartLoading(false);
   };
 
   const addCartItem = async (product: Product) => {
@@ -52,6 +46,11 @@ const useCartService = () => {
   const updateCartItemQuantity =
     (cartId: string) => async (quantity: number) => {
       if (!cartId) return;
+
+      if (quantity === 0) {
+        deleteCartItem(cartId);
+        return;
+      }
 
       const response = await fetch(`${cartItemsUrl}/${cartId}`, {
         method: 'PATCH',
@@ -81,6 +80,8 @@ const useCartService = () => {
     };
 
   const deleteCartItem = async (cartId: string) => {
+    if (!cartId) return;
+
     const response = await fetch(`${cartItemsUrl}/${cartId}`, {
       method: 'DELETE',
       headers: {
